+++ src/error/_reexport_std.rs
@@ -1,13 +1,13 @@
-// devela/src/code/error/_reexport_std.rs
+// devela/src/error/_reexport_std.rs
 
 #[allow(unused_imports, reason = "symlinked from devela")]
 use crate::{_reexport, _tags};
 
-_reexport! { rust: std::backtrace, location: "code/error", tag: _tags!(debug),
+_reexport! { rust: std::backtrace, location: "error", tag: _tags!(debug),
     doc: "A captured OS thread stack backtrace.",
     Backtrace
 }
-_reexport! { rust: std::backtrace, location: "code/error", tag: _tags!(debug),
+_reexport! { rust: std::backtrace, location: "error", tag: _tags!(debug),
     doc: "The current status of a backtrace.",
     BacktraceStatus
 }