+++ src/error/kind.rs
@@ -1,6 +1,6 @@
-// devela/src/code/error/errors.rs
+// devela/src/error/kind.rs
 //
-//! Defines the most general error types.
+//! Reusable failure categories.
 //
 // TOC
 // - individual error types:
@@ -16,23 +16,23 @@ use crate::{_TAG_NO, _TAG_VALUE, define_error};
 /* individual errors */
 
 define_error![individual: pub struct FailedErrorConversion;
-    #[derive(Default)], +location: "code/error", +tag: _TAG_VALUE!(),
+    #[derive(Default)], +location: "error/code", +tag: _TAG_VALUE!(),
     DOC_FAILED_CONVERSION = "A failed conversion between two error types.",
     self+f => write!(f, "Failed to convert between error types."),
 ];
 define_error![individual: pub struct NotImplemented;
-    #[derive(Default)], +location: "code/error", +tag: _TAG_NO!(),
+    #[derive(Default)], +location: "error/code", +tag: _TAG_NO!(),
     DOC_NOT_IMPLEMENTED = "The requested functionality is not implemented.",
     self+f => write!(f, "The requested functionality is not implemented."),
 ];
 define_error![individual: pub struct NotSupported;
-    #[derive(Default)], +location: "code/error", +tag: _TAG_NO!(),
+    #[derive(Default)], +location: "error/code", +tag: _TAG_NO!(),
     DOC_NOT_SUPPORTED = "The requested functionality is not supported by this type.",
     self+f => write!(f, "The requested functionality is not supported by this type."),
 ];
 
 define_error![individual: pub struct InvalidValue;
-    #[derive(Default)], +location: "code/error", +tag: _TAG_VALUE!(),
+    #[derive(Default)], +location: "error/code", +tag: _TAG_VALUE!(),
     DOC_INVALID_VALUE = "An invalid value was received for the given type or operation.",
     self+f => write!(f, "An invalid value was received for the given type or operation."),
 ];
@@ -40,7 +40,7 @@ define_error![individual: pub struct InvalidValue;
 /* composite errors */
 
 define_error! { composite: fmt(f)
-    +location: "code/error",
+    +location: "error/code",
     +tag: _TAG_NO!(),
     /// An error composite of [`NotImplemented`] + [`NotSupported`].
     pub enum NotAvailable {